requirements.txt
python-dotenv
tweepy
TwitterAPI